+++ api/quotes/delete.php
@@ -10,17 +10,32 @@ $db = $database->connect();
 //Create author object
 $quote = new Quote($db);
 
-//check for id
-$quote->id = isset($_GET['id']) ? $_GET['id'] : die();
+$data = json_decode(file_get_contents("php://input"));
 
+//check for id
+if(!isset($data->id)){
+   echo json_encode(
+        array('message' => 'Missing Required Parameters')
+    );
+    return;
+}
+
+$quote->id = $data->id;
+
+if (!$quote->read_single()) {
+    echo json_encode(
+        array('message' => 'No Quotes Found')
+    );
+    return;
+}
 
 if($quote->delete()) {
-    http_response_code(204);
-    echo json_encode(
-        array('message' => 'Successfully deleted')
-    );
+    echo json_encode(array('id'=>$data->id
+    ));
+         
 } else {
-    http_response_code(404);
-    echo json_encode(array('message' => 'Quote with ID ' . $quote->id . ' not found.'));
+    echo json_encode(
+      array('message' => 'No Quotes Found'));
 }
+
